Finish setup command
