+++ src/cli/library_upload.js
@@ -1,7 +1,7 @@
 import { LibraryContributeCommand, LibraryContributeCommandSite } from '../cmd';
 import { convertApiError } from '../cmd/api';
 import chalk from 'chalk';
-import log from '../app/log';
+import log from '../lib/log';
 import { spin } from '../app/ui';
 import { buildAPIClient } from './apiclient';
