+++ src/cmd/api.js
@@ -1,7 +1,7 @@
 import fs from 'fs';
 import when from 'when';
 import Particle from 'particle-api-js';
-import log from '../app/log';
+import log from '../lib/log';
 import _ from 'lodash';
 import url from 'url';
 import chalk from 'chalk';
